Fixed task 2 error

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,5 @@
 #include "function_pointers.h"
-#include <stdlib.h>
+
 /*
  * array_iterator - executes a function given on each element of an array.
  *
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned long int i = 0;
+	unsigned int i = 0;
 
 	if (array == NULL || action == NULL)
 		return;
